files/home/.claude/skills/dspy-optimizer/scripts/pyproject.toml
[project]
name = "dspy-optimizer-scripts"
version = "0.1.0"
description = "Portable DSPy optimization pipeline scripts (GEPA default)"
requires-python = ">=3.10"
dependencies = [
    "dspy>=3.1.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
knn = ["openai>=1.0", "numpy>=1.24"]
